refactor: organize top-rated feature files
