+++ data/src/main/java/com/london/data/di/NetworkModule.kt
@@ -3,9 +3,10 @@ package com.london.data.di
 import android.content.Context
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.london.data.BuildConfig
-import com.london.data.remote.interceptor.AuthInterceptor
 import com.london.data.local.preference.AuthPreferences
 import com.london.data.local.preference.SharedPrefsTokenProvider
+import com.london.data.local.source.device.DeviceConfigurationDataSource
+import com.london.data.remote.interceptor.AuthInterceptor
 import com.london.data.remote.service.auth.AuthApiService
 import com.london.data.remote.service.details.actor.ActorDetailsApiService
 import com.london.data.remote.service.details.movie.MovieDetailsApiService
@@ -13,9 +14,8 @@ import com.london.data.remote.service.details.tvshow.TvShowDetailsApiService
 import com.london.data.remote.service.home.PopularApiService
 import com.london.data.remote.service.reviews.ReviewsApiService
 import com.london.data.remote.service.search.SearchApiService
-import com.london.data.remote.service.toprated.movie.TopRatedMovieApiService
-import com.london.data.remote.service.toprated.tvseries.TopRatedTvSeriesApiService
-import com.london.data.local.source.device.DeviceConfigurationDataSource
+import com.london.data.remote.service.toprated.TopRatedMovieApiService
+import com.london.data.remote.service.toprated.TopRatedTvSeriesApiService
 import com.london.domain.repository.SessionTokenProvider
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
