+++ data/src/main/java/com/london/data/mapper/toprated/TopRatedMovieMapper.kt
@@ -1,6 +1,6 @@
 package com.london.data.mapper.toprated
 
-import com.london.data.remote.model.toprated.movie.model.TopRatedMovieRemote
+import com.london.data.remote.model.toprated.TopRatedMovieRemote
 import com.london.data.utils.asImageUrlOrEmpty
 import com.london.data.utils.isTrue
 import com.london.data.utils.orZero
