+++ data/src/main/java/com/london/data/remote/model/toprated/TopRatedMovieRemote.kt
@@ -1,4 +1,4 @@
-package com.london.data.remote.model.toprated.movie.model
+package com.london.data.remote.model.toprated
 
 
 import kotlinx.serialization.SerialName
