+++ data/src/main/java/com/london/data/remote/service/toprated/TopRatedMovieApiService.kt
@@ -1,9 +1,9 @@
 @file:KoverIgnore
 
-package com.london.data.remote.service.toprated.movie
+package com.london.data.remote.service.toprated
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.toprated.movie.model.TopRatedMovieRemote
+import com.london.data.remote.model.toprated.TopRatedMovieRemote
 import com.london.domain.KoverIgnore
 import retrofit2.Response
 import retrofit2.http.GET
