+++ data/src/main/java/com/london/data/remote/service/toprated/TopRatedTvSeriesApiService.kt
@@ -1,9 +1,9 @@
 @file:KoverIgnore
 
-package com.london.data.remote.service.toprated.tvseries
+package com.london.data.remote.service.toprated
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.toprated.tvshow.model.TopRatedTvSeriesRemote
+import com.london.data.remote.model.toprated.TopRatedTvSeriesRemote
 import com.london.domain.KoverIgnore
 import retrofit2.Response
 import retrofit2.http.GET
