+++ data/src/main/java/com/london/data/remote/source/toprated/movie/TopRatedMovieRemoteDataSource.kt
@@ -3,9 +3,8 @@
 package com.london.data.remote.source.toprated.movie
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.toprated.movie.model.TopRatedMovieRemote
+import com.london.data.remote.model.toprated.TopRatedMovieRemote
 import com.london.domain.KoverIgnore
-import retrofit2.Response
 
 interface TopRatedMovieRemoteDataSource {
     suspend fun getTopRatedMovies(
