+++ data/src/main/java/com/london/data/remote/source/toprated/movie/TopRatedMovieRemoteDataSourceImpl.kt
@@ -3,8 +3,8 @@
 package com.london.data.remote.source.toprated.movie
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.toprated.movie.model.TopRatedMovieRemote
-import com.london.data.remote.service.toprated.movie.TopRatedMovieApiService
+import com.london.data.remote.model.toprated.TopRatedMovieRemote
+import com.london.data.remote.service.toprated.TopRatedMovieApiService
 import com.london.data.remote.source.base.BaseRemoteDatasource
 import com.london.domain.KoverIgnore
 import org.koin.core.annotation.Single
