+++ data/src/main/java/com/london/data/remote/source/toprated/tvseries/TopRatedTvRemoteDataSource.kt
@@ -3,7 +3,7 @@
 package com.london.data.remote.source.toprated.tvseries
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.toprated.tvshow.model.TopRatedTvSeriesRemote
+import com.london.data.remote.model.toprated.TopRatedTvSeriesRemote
 import com.london.domain.KoverIgnore
 
 interface TopRatedTvRemoteDataSource {
