+++ data/src/main/java/com/london/data/remote/source/toprated/tvseries/TopRatedTvRemoteDataSourceImpl.kt
@@ -3,8 +3,8 @@
 package com.london.data.remote.source.toprated.tvseries
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.toprated.tvshow.model.TopRatedTvSeriesRemote
-import com.london.data.remote.service.toprated.tvseries.TopRatedTvSeriesApiService
+import com.london.data.remote.model.toprated.TopRatedTvSeriesRemote
+import com.london.data.remote.service.toprated.TopRatedTvSeriesApiService
 import com.london.data.remote.source.base.BaseRemoteDatasource
 import com.london.domain.KoverIgnore
 import org.koin.core.annotation.Single
