+++ domain/src/main/java/com/london/domain/repository/toprated/TopRatedTvSeriesRepository.kt
@@ -1,4 +1,4 @@
-package com.london.domain.repository
+package com.london.domain.repository.toprated
 
 import com.london.domain.entity.PagedFetchResponse
 import com.london.domain.entity.toprated.TopRatedTvSeries
