+++ domain/src/main/java/com/london/domain/usecase/toprated/GetTopRatedTvSeriesUseCase.kt
@@ -1,8 +1,8 @@
-package com.london.domain.usecase
+package com.london.domain.usecase.toprated
 
 import com.london.domain.entity.PagedFetchResponse
 import com.london.domain.entity.toprated.TopRatedTvSeries
-import com.london.domain.repository.TopRatedTvSeriesRepository
+import com.london.domain.repository.toprated.TopRatedTvSeriesRepository
 import org.koin.core.annotation.Provided
 import org.koin.core.annotation.Single
